refactor Loading component into the ui directory
