+++ src/components/Archive/SoldItem/index.js
@@ -4,7 +4,7 @@ import { Link, useParams } from 'react-router-dom'
 import moment from 'moment'
 import DeleteSoldItemModal from '../../ui/Modals/DeleteSoldItemModal'
 import Unauthorized from '../../ui/Alerts/Unauthorized'
-import Loading from '../../Spinners/Loading'
+import Loading from '../../ui/Spinners/Loading'
 import { SendGetRequest, SendPostRequest } from '../../../helpers/SendApiRequest'
 import classes from './index.module.css'
