+++ src/components/Blog/Categories/index.js
@@ -1,7 +1,7 @@
 import { React, useEffect, useState } from 'react'
 import { useSelector } from 'react-redux'
 import Unauthorized from '../../ui/Alerts/Unauthorized'
-import Loading from '../../Spinners/Loading'
+import Loading from '../../ui/Spinners/Loading'
 import { SendGetRequest } from '../../../helpers/SendApiRequest'
 
 function Categories() {
