+++ src/components/Blog/Post/index.js
@@ -3,7 +3,7 @@ import { useDispatch, useSelector } from 'react-redux'
 import { useParams } from 'react-router-dom'
 import moment from 'moment'
 import Unauthorized from '../../ui/Alerts/Unauthorized'
-import Loading from '../../Spinners/Loading'
+import Loading from '../../ui/Spinners/Loading'
 import { openModal, closeModal } from '../../../features/Modal'
 import { SendGetRequest } from '../../../helpers/SendApiRequest'
 import classes from './index.module.css'
