+++ src/components/Blog/Sidebar/index.js
@@ -2,7 +2,7 @@ import { React, useEffect, useState } from 'react'
 import { useSelector } from 'react-redux'
 import { Link } from 'react-router-dom'
 import Unauthorized from '../../ui/Alerts/Unauthorized'
-import Loading from '../../Spinners/Loading'
+import Loading from '../../ui/Spinners/Loading'
 import MonthYear from '../MonthYear'
 import Year from '../Year'
 import { SendGetRequest } from '../../../helpers/SendApiRequest'
