+++ src/pages/PostsByUser.js
@@ -2,7 +2,7 @@ import { React, useEffect, useState } from 'react'
 import { useSelector } from 'react-redux'
 import { useParams } from 'react-router-dom'
 import Unauthorized from '../components/ui/Alerts/Unauthorized'
-import Loading from '../components/Spinners/Loading'
+import Loading from '../components/ui/Spinners/Loading'
 import Posts from '../components/Blog/Posts'
 import { SendGetRequest } from '../helpers/SendApiRequest'
 import classes from './PostsByUser.module.css'
